hwc: FPS property to show compositor performance
Use setprop debug.hwc.showfps 1 to show the average frames per second
displayed over the last 8 compositions.

debug.hwc.showfps can be set dynamically, the overhead of accessing Android
properties should be minor i.e. an access to an mmap'd() file, so this
should not be a significant overhead to composition when disabled.

Reworked from p-ics-mr1:
9ea53ce HWC: debug.hwc.showfps Android property shows current FPS
66a531d hwc: enable debug.hwc.showfps to be set dynamically

Change-Id: Ie2303e6ae9e73f38cf563d6ed6079928054e4c16
Signed-off-by: Tony Lofthouse <a0741364@ti.com>